+++ app/application/representers/openstruct_with_links.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module TranSound
-  module Response
+  module Representer
     # OpenStruct for deserializing json with hypermedia
     class OpenStructWithLinks < OpenStruct
       attr_accessor :links
