+++ app/application/representers/shows_representer.rb
@@ -3,37 +3,17 @@
 require 'roar/decorator'
 require 'roar/json'
 
-# Represents essential Repo information for API output
+require_relative 'openstruct_with_links'
+require_relative 'show_representer'
+
 module TranSound
   module Representer
-    # Represent a show entity as Json
-    class Show < Roar::Decorator
+    # Represents list of shows for API output
+    class ShowsList < Roar::Decorator
       include Roar::JSON
-      include Roar::Hypermedia
-      include Roar::Decorator::HypermediaConsumer
 
-      property :id
-      property :origin_id
-      property :description
-      property :images
-      property :name
-      property :publisher
-      property :type
-      property :show_url
-
-      link :self do
-        "#{ENV.fetch('API_HOST', nil)}/api/v1/podcast_info/#{show_type}/#{show_id}"
-      end
-
-      private
-
-      def show_type
-        represented.type
-      end
-
-      def show_id
-        represented.id
-      end
+      collection :shows, extend: Representer::Show,
+                         class: Representer::OpenStructWithLinks
     end
   end
 end
