+++ config/secrets_example.yml
@@ -1,20 +1,21 @@
 ---
 development:
+  # Infrastructure
   DB_FILENAME: db/local/dev.db
   spotify_Client_ID: provision_spotify_api_ID
   spotify_Client_secret: provision_spotify_api_secret
-  spotify_gettoken_time: provision_gettoken_time
-  spotify_temp_token: provision_spotify_api_temp_token
-
+  # Application
+  API_HOST: http://localhost:9292
+  APP_HOST: http://localhost:9292
 test:
+  # Infrastructure
   DB_FILENAME: db/local/test.db
   spotify_Client_ID: provision_spotify_api_ID
   spotify_Client_secret: provision_spotify_api_secret
-  spotify_gettoken_time: provision_gettoken_time
-  spotify_temp_token: provision_spotify_api_temp_token
-
+  # Application
+  API_HOST: http://localhost:9292
+  APP_HOST: http://localhost:9292
 production:
   spotify_Client_ID: provision_spotify_api_ID
   spotify_Client_secret: provision_spotify_api_secret
-  spotify_gettoken_time: provision_gettoken_time
-  spotify_temp_token: provision_spotify_api_temp_token
+
