+++ config/temp_token.yml
@@ -1,3 +1,3 @@
 ---
-spotify_gettoken_time: '20231211134141'
-spotify_temp_token: BQAFzzjMIfOLmIphu35AmMBtityMAmoJpYlPm53yLj4lV8NMUhllugMAMFRCzQV478_jogjGC3_QmAen99GUIWC3PEsDId9wGJLhbNCKIcrJoETlHWY
+spotify_gettoken_time: '20231211175308'
+spotify_temp_token: BQBWG5dKZOTMNDeZS0As-obGHtiYRixiL9tq2hW_ePKuxBkf6p37dTu59uehkcNhPCYcBAQ0FdWnVKMqtHFcCGaBHRI4k8i9RSdfiCk7hnfdvJrnysM
